--add android:usesCleartextTraffic="true" in Application Tag of Manifest file
Indicates whether the app intends to use cleartext network traffic, such as cleartext HTTP. The default value is "true". When the attribute is set to "false", platform components (for example, HTTP and FTP stacks, DownloadManager, MediaPlayer) will refuse the app's requests to use cleartext traffic. Third-party libraries are strongly encouraged to honor this setting as well. The key reason for avoiding cleartext traffic is the lack of confidentiality, authenticity, and protections against tampering: a network attacker can eavesdrop on transmitted data and also modify it without being detected.